dom/payments/test/test_cleanupPayment.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1408234
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1408234</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript">

  "use strict";
  SimpleTest.waitForExplicitFinish();

  var gUrl = SimpleTest.getTestFileURL('CleanupChromeScript.js');
  var gScript = SpecialPowers.loadChromeScript(gUrl);
  var testName = "";

  function testFailHandler(message) {
    ok(false, testName + ": " + message);
  }
  function testPassHandler(message) {
    ok(true, testName + ": " + message);
  }
  gScript.addMessageListener("test-fail", testFailHandler);
  gScript.addMessageListener("test-pass", testPassHandler);
  
  function cleanupCheck(resolveFunc) {
    gScript.addMessageListener("cleanup-check-complete", function checkCompleteHandler() {
      gScript.removeMessageListener("cleanup-check-complete", checkCompleteHandler);
      resolveFunc();
    });
    gScript.sendAsyncMessage("cleanup-check");
  }

  function paymentNumCheck(resolveFunc, numPayments) {
    gScript.addMessageListener("payment-num-check-complete", function checkCompleteHandler() {
      gScript.removeMessageListener("payment-num-check-complete", checkCompleteHandler);
      resolveFunc();
    });
    gScript.sendAsyncMessage("payment-num-check", numPayments);
  }

  function paymentNumSet() {
    gScript.addMessageListener("payment-num-set-complete", function setCompleteHandler() {
      gScript.removeMessageListener("payment-num-set-complete", setCompleteHandler);
    });
    gScript.sendAsyncMessage("payment-num-set");
  }

  function testCleanupByReloading() {
    paymentNumSet();
    return new Promise((resolve, reject) => {
      testName = "testCleanupByReloading";
      let expectedResults = ["successful", "successful"];
      let nextStatus = ["creating", "reloading"];
      let currStatus = nextStatus.shift();
      let ifr = document.createElement('iframe');
      
      let listener = function(event) {
        let expected = expectedResults.shift();
        is(event.data, expected,
          testName + ": Expected '" + expected + "' when " + currStatus +
          ", but got '" + event.data + "'");
        if (currStatus === "creating") {
          ifr.contentWindow.location.reload();
        } else if (currStatus === "reloading") {
          window.removeEventListener("message", listener);
          paymentNumCheck(resolve, 1);
          document.body.removeChild(ifr);
        }
        currStatus = nextStatus.shift();
      }
      window.addEventListener("message", listener);
      ifr.src = "simple_payment_request.html";
      document.body.appendChild(ifr);
    });
  }

  function testCleanupByRedirecting() {
    return new Promise((resolve, reject) => {
      testName = "testCleanupByRedirecting";
      let expectedResults = ["successful", "successful"];
      let nextStatus = ["creating", "redirecting"];
      let currStatus = nextStatus.shift();
      let ifr = document.createElement('iframe');
      let listener = function(event) {
        let expected = expectedResults.shift();
        is(event.data, expected,
          testName + ": Expected '" + expected + "' when " + currStatus +
          ", but got '" + event.data + "'");
        if (currStatus === "creating") {
          ifr.src = "blank_page.html";
        } else if (currStatus === "redirecting"){
          window.removeEventListener("message", listener);
          cleanupCheck(resolve);
          document.body.removeChild(ifr);
        }
        currStatus = nextStatus.shift();
      };
      window.addEventListener("message", listener);
      ifr.src = "simple_payment_request.html";
      document.body.appendChild(ifr);
    });
  }

  function testCleanupByRedirectingAfterShow() {
    return new Promise((resolve, reject) => {
      testName = "testCleanupByRedirectingAfterShow";
      let nextStatus = ["creating", "showing", "redirecting"];
      let currStatus = nextStatus.shift();
      let expectedResults = ["successful", "successful", "successful"];
      let ifr = document.createElement('iframe');
      let listener = (event) => {
        let expected = expectedResults.shift();
        is(event.data, expected,
          testName + ": Expected '" + expected + "' when " + currStatus +
          ", but got '" + event.data + "'");

        if (currStatus === "creating") {
          ifr.contentWindow.postMessage("show PaymentRequest", "*");
        } else if (currStatus === "showing") {
          ifr.src = "blank_page.html";
        } else if (currStatus === "redirecting") {
          window.removeEventListener("message", listener);
          cleanupCheck(resolve);
          document.body.removeChild(ifr);
        }
        currStatus = nextStatus.shift();
      }
      window.addEventListener("message", listener);
      ifr.src = "simple_payment_request.html";
      document.body.appendChild(ifr);
    });
  }

  function testCleanupByRemovingIframe() {
    return new Promise((resolve, reject) => {
      testName = "testCleanupByRemovingIframe";
      let expectedResults = ["successful"];
      let nextStatus = ["creating"];
      let currStatus = nextStatus.shift();
      let ifr = document.createElement('iframe');
      let listener = function(event) {
        let expected = expectedResults.shift();
        is(event.data, expected,
          testName + ": Expected '" + expected + "' when " + currStatus +
          ", but got '" + event.data + "'");
        document.body.removeChild(ifr);
        window.removeEventListener("message", listener);
        cleanupCheck(resolve);
      };
      window.addEventListener("message", listener);
      ifr.src = "simple_payment_request.html";
      document.body.appendChild(ifr);
    });
  }

  function teardown() {
    gScript.removeMessageListener("test-fail", testFailHandler);
    gScript.removeMessageListener("test-pass", testPassHandler);
    gScript.destroy();
    SimpleTest.finish();
  }

  function runTests() {
    testCleanupByReloading()
    .then(testCleanupByRedirecting)
    .then(testCleanupByRedirectingAfterShow)
    .then(testCleanupByRemovingIframe)
    .then(teardown)
    .catch( e => {
      ok(false, "Unexpected error: " + e.name);
      SimpleTest.finish();
    });
  }

  window.addEventListener('load', function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
      ]
    }, runTests);
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1408234">Mozilla Bug 1408234</a>
</body>
</html>
